Refactor) redux toolkit 적용
